hw1_http_server/tests/test_requirements.txt
psycopg2==2.9.5
python-dotenv==0.21.1
requests==2.28.2
pytest==6.2.5